requirements.txt
# Core data processing
pandas>=2.0.0
pyarrow>=14.0.1  # Parquetファイルの処理に必要
fastparquet>=2023.8.0  # 代替のParquetエンジン

# Text processing
pyahocorasick>=2.0.0  # 高速な文字列検索

# Utility packages
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2023.3

# Optional but recommended
tqdm>=4.65.0  # プログレスバーの表示